Fix whole bunch of bug

+++ src/main/java/com/application/controllers/FileLinkRelController.java
@@ -20,18 +20,20 @@ public class FileLinkRelController {
 	private FileLinkRefRepository fileLinkRefRepository;
 	@Autowired
 	private FileTypeRepository fileTypeRepository;
+
 	@Autowired
 	private MinioStorageService minioStorageService;
 
 	// ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
-	// DB-V5 OK!
-	// Add image to track.
+	// DB-V6 OK!
+	// NOTE | Add image to track.
 	public String insertNewTrackObjectLinkRel(MultipartFile multipartFile, int typeId, int recordRel) {
 		FileLinkRefModel newFileRecord = new FileLinkRefModel();
 
 		newFileRecord.setFileType(fileTypeRepository.findById(typeId)
 				.orElseThrow(() -> new ExceptionFoundation(EXCEPTION_CODES.FILE_INVALID_TYPE, HttpStatus.I_AM_A_TEAPOT,
 						"[ FILE_INVALID_TYPE ] A file type with this ID does not exists.")));
+
 		newFileRecord.setFileId(minioStorageService.uploadImageToStorage(multipartFile, typeId + "-",
 				newFileRecord.getFileType().getPathRel()));
 		newFileRecord.setTargetRef(recordRel);
@@ -41,7 +43,7 @@ public class FileLinkRelController {
 
 	// ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
 	// DB-V6 OK!
-	// Upload track to storage
+	// NOTE | Upload track to storage
 	public String uploadNewTrack(MultipartFile trackFile, int trackId) {
 		String fileName = minioStorageService.uploadTrackToStorage(trackFile, "tracks/musics/");
 		FileLinkRefModel newFile = new FileLinkRefModel();
@@ -53,56 +55,41 @@ public class FileLinkRelController {
 		fileLinkRefRepository.save(newFile);
 		return fileName;
 	}
-	
 
 	// ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
 	// DB-V5 OK!
 	// Delete file by its name.
+	// CONDITION | If not found, just ignore.
+	// CONDITION | If found record but nit in Minio, delete record.
 	public void deleteTargetFileByName(String fileName) {
-		FileLinkRefModel target = fileLinkRefRepository.findById(fileName).orElseThrow(() -> new ExceptionFoundation(
-				EXCEPTION_CODES.SEARCH_NOT_FOUND, HttpStatus.NOT_FOUND,
-				"[ FileLinkRelController] Can't delete this record because the file " + fileName + " is unreachable."));
-		String destination = target.getFileType().getPathRel() + fileName;
-		minioStorageService.DeleteObjectFromMinIoByPathAndName(destination);
-		fileLinkRefRepository.delete(target);
+		FileLinkRefModel target = fileLinkRefRepository.findById(fileName).orElse(null);
+		if (target != null) {
+			String destination = target.getFileType().getPathRel() + fileName;
+			try {
+				minioStorageService.DeleteObjectFromMinIoByPathAndName(destination);
+			} catch (Exception exc) {
+				System.out.println(
+						"[ MINIO_OBJECT_UNREACHABLE ] This is not an error! An object not found in the storage, the record will be deleted."
+								+ "\n" + exc.getLocalizedMessage());
+			}
+			fileLinkRefRepository.delete(target);
+		} else {
+			return;
+		}
+
 	}
 
 	// ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
 	// DB-V6 OK!
-	// Upload track to storage
-	public String deleteTrack(String fileName) {
-		return "";
-	}
-
-	// ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
-	// DB-V5 OK!
-	// Delete track image using type ID and ref ID.
-	public void deleteTargetFileByTypeIdAndLinkRef(int typeId, int refId) {
-		FileLinkRefModel target = fileLinkRefRepository.findByTargetRefAndTypeId(typeId, refId);
-		String destination = target.getFileType().getPathRel() + target.getFileId();
-		minioStorageService.DeleteObjectFromMinIoByPathAndName(destination);
-		fileLinkRefRepository.delete(target);
-	}
-
-	// ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
-	// DB-V5 OK!
 	// RetriveImageByName
+	// NOTE | It there is a record, an image will be grabed and sent to you.
+	// EXCEPTION | 1005 | MINIO_OBJECT_DOES_NOT_EXIST
 	public Resource retriveImageByName(String fileName) {
-		FileLinkRefModel targetFile = fileLinkRefRepository.findById(fileName)
-				.orElseThrow(() -> new ExceptionFoundation(EXCEPTION_CODES.SEARCH_NOT_FOUND, HttpStatus.NOT_FOUND,
+		FileLinkRefModel targetFile = fileLinkRefRepository.findById(fileName).orElseThrow(
+				() -> new ExceptionFoundation(EXCEPTION_CODES.MINIO_OBJECT_DOES_NOT_EXIST, HttpStatus.NOT_FOUND,
 						"[ FileLinkRelController ] File name " + fileName + "does not exist in the record."));
 		String fileNameLocation = targetFile.getFileType().getPathRel() + fileName;
-
 		return minioStorageService.getImageFromMinIoByNameLocation(fileNameLocation);
-	}
-
-	// ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
-	// DB-V5 OK!
-	// RetriveImageByTargetRef
-	public Resource retriveImageByTargetRef(int typeId, int targetRef) {
-		FileLinkRefModel fileModel = fileLinkRefRepository.findByTargetRefAndTypeId(typeId, targetRef);
-		String targetFileInMinIo = fileModel.getFileType().getPathRel() + fileModel.getFileId();
-		return minioStorageService.getImageFromMinIoByNameLocation(targetFileInMinIo);
 	}
 
 	// ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
@@ -113,6 +100,40 @@ public class FileLinkRelController {
 			return true;
 		} else {
 			return false;
+		}
+	}
+
+	// NOT IN USE
+	// ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
+	// DB-V6 OK!
+	// RetriveImageByTargetRef
+	// NOTE | Instead of looking by their name, this method looks into their typeId
+	// and Ref, which are their identity, to pull out an image/
+	public Resource retriveImageByTargetRef(int typeId, int targetRef) {
+		FileLinkRefModel fileModel = fileLinkRefRepository.findByTargetRefAndTypeId(typeId, targetRef);
+		String targetFileInMinIo = fileModel.getFileType().getPathRel() + fileModel.getFileId();
+		return minioStorageService.getImageFromMinIoByNameLocation(targetFileInMinIo);
+	}
+
+	// NOT IN USE
+	// ▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬▬
+	// DB-V6 OK!
+	// NOTE | Delete track image using type ID and ref ID.
+	public void deleteTargetFileByTypeIdAndLinkRef(int typeId, int refId) {
+		FileLinkRefModel target = fileLinkRefRepository.findByTargetRefAndTypeId(typeId, refId);
+		if (target != null) {
+			String destination = target.getFileType().getPathRel() + target.getFileId();
+			try {
+				minioStorageService.DeleteObjectFromMinIoByPathAndName(destination);
+				fileLinkRefRepository.delete(target);
+			} catch (Exception exc) {
+				System.out.println(
+						"[ MINIO_OBJECT_UNREACHABLE ] This is not an error! An object not found in the storage, the record will be deleted."
+								+ "\n" + exc.getLocalizedMessage());
+			}
+		} else {
+			System.out.println(
+					"[ MINIO_OBJECT_UNREACHABLE ] This is not an error! An object not found in the storage, the record will be deleted.");
 		}
 	}
